pyproject.toml
[tool.poetry]
name = "datatricks"
version = "0.1.0"
description = "Common Functions for Data Science/Analysis"
authors = ["Shivam Anand <shivam01anand@gmail.com>"]
keywords = ["data", "analytics","datascience","pandas","etl"]
license = "MIT"
homepage = "https://github.com/MattiooFR/package_name"
repository = "https://github.com/MattiooFR/package_name"
classifiers = ["Topic :: Software Development :: Build Tools",
               "Topic :: Software Development :: Libraries :: Python Modules",
               "Framework :: Jupyter"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.4.1"
Pillow = "^9.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
